Move Xdr logic under history/xdr package.
