+++ history/src/main/java/com/groupon/lex/metrics/history/xdr/support/XdrStreamIterator.java
@@ -11,7 +11,6 @@ import static java.util.Objects.requireNonNull;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import org.acplt.oncrpc.OncRpcException;
-import org.dcache.xdr.XdrBufferDecodingStream;
 
 /**
  *
