+++ history/src/test/java/com/groupon/lex/metrics/history/xdr/XdrReadBackTest.java
@@ -18,8 +18,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Stream;
-import org.dcache.xdr.XdrBufferDecodingStream;
-import org.dcache.xdr.XdrBufferEncodingStream;
+import com.groupon.lex.metrics.history.xdr.support.XdrBufferDecodingStream;
+import com.groupon.lex.metrics.history.xdr.support.XdrBufferEncodingStream;
 import static org.junit.Assert.assertEquals;
 import org.joda.time.DateTime;
 import org.joda.time.DateTimeZone;
